BankApplication_TDD/src/com/capgemini/test/AccountServiceImplTest.java
package com.capgemini.test;

import static org.junit.Assert.*;

import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import com.capgemini.beans.Account;
import com.capgemini.exceptions.InsufficientBalanceException;
import com.capgemini.exceptions.InsufficientInitialAmountException;
import com.capgemini.exceptions.InvalidAccountNumberException;
import com.capgemini.repository.AccountRepository;
import com.capgemini.service.AccountService;
import com.capgemini.service.AccountServiceImpl;
import static org.mockito.Mockito.when;

public class AccountServiceImplTest {

	@Mock
	AccountRepository accountRepository;
	
	AccountService accountService;
	
	@Before
	public void setUp() throws Exception{
		MockitoAnnotations.initMocks(this);
		accountService = new AccountServiceImpl(accountRepository);
	}
	
	@Test(expected=com.capgemini.exceptions.InsufficientInitialAmountException.class)
	public void whenTheAmountIsLessThanFiveHundredSystemShouldThrowException() throws InsufficientInitialAmountException {		
		accountService.createAccount(101, 400);
	}
	
	@Test
	public void whenTheValidInfoIsPassedAccountShouldCreateSuccessfully() throws InsufficientInitialAmountException, InvalidAccountNumberException{
		Account account = new Account();
		account.setAccountNumber(101);
		account.setAmount(5000);
		when(accountRepository.save(account)).thenReturn(true);
		assertEquals(account, accountService.createAccount(101,5000));
	}
	
	@Test(expected=com.capgemini.exceptions.InvalidAccountNumberException.class)
	public void whenTheAccountNumberIsInValid() throws InvalidAccountNumberException, InsufficientInitialAmountException{
		accountService.depositAmount(-1, 500);
		accountService.createAccount(-1, 500);
	}

	@Test(expected=com.capgemini.exceptions.InsufficientBalanceException.class)
	public void whenTheBalanceIsInsufficient() throws InsufficientBalanceException{
		accountService.withdrawAmount(1111,-12);
	}

	@Test
	public void whenTheValidInfoIsPassedAmountDepositedSuccessfully() throws InvalidAccountNumberException, InsufficientBalanceException{
		Account account = new Account();
		account.setAccountNumber(101);
		account.setAmount(5000);
		when(accountRepository.save(account)).thenReturn(true);
		assertEquals(account, accountService.depositAmount(101,5000));
	}
}
